+++ src/hooks/api/auth/useSendCodeMutation.ts
@@ -1,0 +1,23 @@
+import { useMutation } from "@tanstack/react-query";
+import { MSW_BASE_URL } from "@/constants/url-constants";
+import type { AxiosError, AxiosResponse } from "axios";
+import { api } from "@/libs/axios";
+import type { SignUpSendRequest } from "@/types/api-request-types/auth-request-types";
+import type {
+  SignUpApiErrorResponse,
+  SignUpSendResponse,
+} from "@/types/api-response-types/auth-response-types";
+
+export function useSendCodeMutation() {
+  return useMutation<SignUpSendResponse, AxiosError<SignUpApiErrorResponse>, SignUpSendRequest>({
+    mutationKey: ["auth", "sendCode"],
+    mutationFn: (payload) =>
+      api
+        .post<
+          SignUpSendResponse,
+          AxiosResponse<SignUpSendResponse>,
+          SignUpSendRequest
+        >(`${MSW_BASE_URL}/users/signup/send/`, payload)
+        .then((res) => res.data),
+  });
+}
