+++ src/hooks/api/auth/useSignupMutation.ts
@@ -1,0 +1,23 @@
+import { useMutation } from "@tanstack/react-query";
+import { MSW_BASE_URL } from "@/constants/url-constants";
+import type { AxiosError, AxiosResponse } from "axios";
+import { api } from "@/libs/axios";
+import type {
+  SignUpApiErrorResponse,
+  SignUpResponse,
+} from "@/types/api-response-types/auth-response-types";
+import type { SignUpRequest } from "@/types/api-request-types/auth-request-types";
+
+export function useSignupMutation() {
+  return useMutation<SignUpResponse, AxiosError<SignUpApiErrorResponse>, SignUpRequest>({
+    mutationKey: ["auth", "signup"],
+    mutationFn: (payload) =>
+      api
+        .post<
+          SignUpResponse,
+          AxiosResponse<SignUpResponse>,
+          SignUpRequest
+        >(`${MSW_BASE_URL}/users/signup`, payload)
+        .then((res) => res.data),
+  });
+}
